requirements.txt
# /ai/requirements.txt
# Python dependencies for AI server with comprehensive plugin support

# Core FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Authentication and security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# HTTP clients and networking
httpx==0.25.2
aiofiles==23.2.1

# Database connections
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Object storage
minio==7.2.0

# AI Model APIs
openai==1.3.0
anthropic==0.7.8

# Rate limiting and performance
slowapi==0.1.9

# Logging and monitoring
loguru==0.7.2

# Utility libraries
python-dateutil==2.8.2
typing-extensions==4.8.0